requirements.txt
altgraph==0.17
appdirs==1.4.4
backcall==0.2.0
black==20.8b1
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
ffmpeg==1.4
future==0.18.2
helpdev==0.7.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
llvmlite==0.34.0
mypy-extensions==0.4.3
numba==0.51.0
numpy==1.21.4
pandas==1.3.4
parso==0.8.1
pathspec==0.8.1
pefile==2019.4.18
pi==0.1.2
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.17
Pygments==2.8.1
PyOpenGL==3.1.5
pyparsing==2.4.7
pyperclip==1.8.0
pypiwin32==223
pyqtgraph==0.11.0
PySide2==5.15.2
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pyzmq==22.0.3
QDarkStyle==2.8.1
QtPy==1.9.0
regex==2020.11.13
scipy==1.5.2
shiboken2==5.15.2
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
